requirements.txt
discord.py>=2.4.0,<3
openai>=1.40.0
python-dotenv>=1.0.0
